Correct buffer sizing advice
The worst case input for rbsp_to_nal is an all-zero byte string. For every two
input bytes there will be an extra 0x03 in the output. An additional byte is
required for the reserved NAL header. Multiplying with 3/2 and rounding up takes
care of both the case when the input has odd and when it has even length.