h264_stream: add multiple bounds checks, & fix double-free of sps_subset_table
There are a bunch of array writes that take their index from the bitstream
but do not check to make sure the index is in-bound.  In many of these
cases, you'd expect the extra values to exist in the stream, so we still
must read them.

When fuzzing h256_analize with AFL++, these missing bounds checks frequently
cause segfaults & other issues.

Tied up in this, there is definitely a double free wrt sps_subset_table &
sps_subset, since the existing code was memcpy'ing pointers to arrays
between structures.  This code copies the contents of the arrays instead.